+++ .gitignore
@@ -233,5 +233,5 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 /.idea/*
-/src/main/java/games/project/statistiques/client64Bit/
+/src/main/java/games/project/modules/statistiques/client64Bit/
 /src/main/PythonFileStats/test/client64Bit/
