+++ src/main/java/games/project/modules/statistiques/views/ViewGamesStats.java
@@ -1,0 +1,4 @@
+package games.project.modules.statistiques.views;
+
+public class ViewGamesStats {
+}
