+++ src/main/java/games/project/modules/statistiques/views/ViewMain.java
@@ -1,6 +1,6 @@
-package games.project.statistiques.views;
+package games.project.modules.statistiques.views;
 
-import games.project.statistiques.StatsLauncher;
+import games.project.modules.statistiques.StatsLauncher;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.ContentDisplay;
