--- src/main/java/games/project/statistiques/views/ViewGamesStats.java
@@ -1,4 +1,0 @@
-package games.project.statistiques.views;
-
-public class ViewGamesStats {
-}
