+++ src/main/ressources/games/project/modules/statistiques/textures/menuScreen.png
